project/pyproject.toml
[tool.poetry]
name = "project"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.38.0"
openai = "^1.43.0"
transformers = {extras = ["torch"], version = "^4.44.2"}
datasets = "^2.21.0"
langchain = "^0.2.16"
matplotlib = "^3.9.2"
langchain-community = "^0.2.16"
python-dotenv = "^1.0.1"
torch = "^2.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
